% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_index.R
\name{A_index}
\alias{A_index}
\title{Species Profile Index A}
\usage{
A_index(x, y)
}
\arguments{
\item{x}{A numeric vector representing tree heights.}

\item{y}{A character or factor vector representing species corresponding to each tree height.}
}
\value{
A numeric value representing the Species Profile Index A.
}
\description{
This function calculates the Species Profile Index A, which measures species diversity
across three height zones:
\itemize{
\item Zone 1: Trees with heights >= 80\% of the maximum height.
\item Zone 2: Trees with heights between 50\% and 80\% of the maximum height.
\item Zone 3: Trees with heights < 50\% of the maximum height.
}
}
\details{
#' For each zone, the species proportions are used to compute a diversity profile index:
\deqn{PI = \sum_{i=1}^{S} |p_i \cdot \ln(p_i)|}

Where:
\itemize{
\item \eqn{S} is the total number of species in the zone.
\item \eqn{p_i} is the proportion of the \eqn{i}-th species in the zone.
}

The Species Profile Index A is the sum of the profile indices across all three zones:
\deqn{A = PI_1 + PI_2 + PI_3}
}
\examples{
x <- c(10, 20, 30, 40, 50, 25, 30, 22, 11, 10)
y <- c(
  "P. patula", "P. patula", "P. patula", "A. religiosa", "Q. castanea",
  "P. patula", "P. patula", "P. patula", "P. patula", "P. patula"
)
A_index(x, y)

}
\references{
Pretzsch, H. (2009). Forest Dynamics, Growth and Yield: From Measurement to Model. Springer-Verlag. https://doi.org/10.1007/978-3-540-88307-4
}
