% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_max.R
\name{A_max}
\alias{A_max}
\title{Maximum Value for the Species Profile Index A}
\usage{
A_max(x)
}
\arguments{
\item{x}{A character or factor vector representing species.}
}
\value{
A numeric value representing the maximum possible Species Profile Index A.
}
\description{
This function calculates the maximum possible value of the Species Profile Index A,
based on the number of distinct species present. It assumes that each species is equally distributed
across the three height zones.
}
\examples{
x <- c(
  "P. patula", "P. patula", "P. patula", "A. religiosa", "Q. castanea",
  "P. patula", "P. patula", "P. patula", "P. patula", "P. patula"
)
A_max(x)

}
