% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bal.R
\name{bal}
\alias{bal}
\title{Calculate Basal Area Larger Index (BAL)}
\usage{
bal(x, y)
}
\arguments{
\item{x}{Numeric vector. Diameters of the subject trees in cm.}

\item{y}{Numeric. Plot area in square meters.}
}
\value{
Numeric vector. The BAL values for each tree in \code{x}.
}
\description{
This function calculates the Basal Area Larger Index (BAL), which is the sum
of the basal area of trees larger than a given subject tree in a specified plot area.
It provides an index of competition based on tree size.
}
\details{
The formula for each tree is:
\deqn{BAL_i = \frac{\sum (BA_j > BA_i)}{A}}
where:
\itemize{
\item \eqn{BA_i} is the basal area of the subject tree.
\item \eqn{BA_j} is the basal area of a competing tree larger than \eqn{BA_i}.
\item \eqn{A} is the plot area in square meters.
}
}
\examples{
x <- c(30, 20, 20, 35, 65) # Tree diameters in cm
bal_result <- bal(x, 400) # Calculate BAL for a plot area of 400 square meters
bal_result

}
