% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca.R
\name{ca}
\alias{ca}
\title{Individual Tree Crown Area}
\usage{
ca(x, y)
}
\arguments{
\item{x}{A numeric variable representing the first tree crown diameter in meters.}

\item{y}{A numeric variable representing the second tree crown diameter in meters.}
}
\value{
A numeric variable representing the tree crown area in square meters.
}
\description{
This function calculates the crown area of an individual tree based on two
measurements of its crown diameters. The crown area is computed using the
average of the two diameters and is assumed to be circular.
}
\examples{
# Calculate the crown area for two crown diameters of 10 m and 15 m
crown_area <- ca(10, 15)
crown_area # Output: Area in square meters

}
