% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cat_d.R
\name{cat_d}
\alias{cat_d}
\title{Diameter Category Classification}
\usage{
cat_d(x)
}
\arguments{
\item{x}{A numeric variable or vector representing tree diameters in centimeters.}
}
\value{
A factor variable or vector representing tree diameter classes in
        increments of 10 cm.
}
\description{
This function categorizes tree diameters into classes of 5 cm, starting
from 2.5 cm. It takes a numeric variable or vector representing tree
diameters and returns the corresponding diameter categories.
}
\examples{
# Create a data frame with tree diameters
d_data <- data.frame(d = c(10.1, 23, 31.5, 40, 52.4, 27, 33, 22, 11, 10))

# Categorize the diameters into classes
d_data$diam_cat <- cat_d(d_data$d)

# Display the categorized data
print(d_data)

}
