% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordxy.R
\name{coordxy}
\alias{coordxy}
\title{Calculate Coordinates from Azimuth and Distance}
\usage{
coordxy(azi, dis, z)
}
\arguments{
\item{azi}{Numeric variable. Azimuth angle in degrees (0-360).}

\item{dis}{Numeric variable. Distance from the reference point (in the same unit as z).}

\item{z}{Numeric variable. Diameter of the circle or side length of the square plot.}
}
\value{
A numeric vector containing the calculated coordinates (x, y).
}
\description{
This function calculates the Cartesian coordinates (x, y) based on a given azimuth
angle and distance from a reference point. The function can be used for both circular
and square plots, where the size of the plot is specified by the diameter (for circles)
or the side length (for squares).
}
\examples{
# Calculate coordinates for an azimuth of 350 degrees and a distance of 15 units
coordxy(350, 15, 20)

}
