% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_dom.R
\name{d_dom}
\alias{d_dom}
\title{Dominant Tree Diameter}
\usage{
d_dom(d, ps)
}
\arguments{
\item{d}{Numeric vector. A vector of tree diameters (in centimeters).}

\item{ps}{Numeric variable. The plot area (in square meters).}
}
\value{
Numeric value representing the mean diameter at breast height
of the dominant trees (in centimeters).
}
\description{
This function calculates the dominant tree diameter, defined as the mean
diameter at breast height (dbh) of the 100 thickest trees within a specified
plot area. It considers the tree diameters provided and determines which
trees qualify as dominant based on their thickness.
}
\examples{
# Example tree diameters and plot area
d <- c(12, 25, 40, 40, 55, 35, 35, 25, 12, 15)
ps <- 400

# Calculate the dominant tree diameter
d_dom(d, ps)
}
