% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dista.R
\name{dista}
\alias{dista}
\title{Compute the Distance between Two Points}
\usage{
dista(x, y, x1, y1)
}
\arguments{
\item{x}{Numeric. The x-coordinate of the first point.}

\item{y}{Numeric. The y-coordinate of the first point.}

\item{x1}{Numeric. The x-coordinate of the second point.}

\item{y1}{Numeric. The y-coordinate of the second point.}
}
\value{
Numeric. The Euclidean distance between the two points.
}
\description{
This function calculates the Euclidean distance between two points in a 2D space.
The distance is computed using the formula:
\deqn{d = \sqrt{(x1 - x)^2 + (y1 - y)^2}}
where \eqn{(x, y)} and \eqn{(x1, y1)} are the coordinates of the two points.
}
\examples{
# Calculate the distance between points (5, 5) and (4, 4)
dista(5, 5, 4, 4)

}
