% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestSDC.R
\docType{package}
\name{forestSDC}
\alias{ForestSDC}
\alias{ForestSDC-package}
\alias{forestSDC}
\title{ForestSDC: Forest stand Structure, tree Diversity and species Composittion.}
\description{
The \code{forestSDC} package is a comprehensive tool set for analyzing data from forest inventories
and permanent research plots. It includes functions to estimate structural indices,
biodiversity metrics, and the nearest neighbor summary statistics.
These tools are essential for assessing forest dynamics, biodiversity, and ecosystem health.
}
\details{
Key features include:
\itemize{
\item Structural indices such as Shannon's diversity, Simpson's complement, and species richness.
\item Spatial analysis using nearest-neighbor methods.
\item Customizable tools for biodiversity and forest dynamics evaluation.
}

The package leverages \code{dplyr} and \code{tidyr} for data manipulation and \code{FNN} for efficient
nearest-neighbor calculations.
}
\author{
Ernesto Rubio \email{ernestorub@gmail.com}
}
