% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Basal Area from Tree Diameter}
\usage{
g(x)
}
\arguments{
\item{x}{Numeric variable. Tree diameter in centimeters.}
}
\value{
Numeric variable. Basal area in square meters.
}
\description{
This function calculates the basal area of a tree based on its diameter at breast height (DBH).
The formula used is \eqn{g = \frac{\pi}{4} \times \left(\frac{d}{100}\right)^2},
where d is the diameter in centimeters.
}
\examples{

x <- 25
g(x) # Calculates the basal area for a tree with a 25 cm diameter.
}
