% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gha.R
\name{Gha}
\alias{Gha}
\title{Calculate Basal Area per Hectare (G/ha)}
\usage{
Gha(d, ps)
}
\arguments{
\item{d}{Numeric vector. A vector of tree diameters (e.g., in cm).}

\item{ps}{Numeric. The plot area (e.g., in square meters).}
}
\value{
Numeric. The basal area per hectare (\eqn{G/ha}).
}
\description{
This function calculates the basal area per hectare (\eqn{G/ha}) using the formula:
\deqn{G/ha = \frac{\sum g}{A} \times 10000}
where:
\itemize{
\item \eqn{\sum g} is the sum of the basal area of individual trees.
\item \eqn{A} is the plot area in square meters.
}
}
\examples{
x <- c(10, 20, 30, 40, 50, 25, 30, 22, 11, 10)
y <- 1000
Gha(x, y)

}
