% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h_dom.R
\name{h_dom}
\alias{h_dom}
\title{Dominant Tree Height (Mean Height of the 100 Thickest Trees in a Hectare)}
\usage{
h_dom(h, d, ps)
}
\arguments{
\item{h}{Numeric vector. Tree heights (in centimeters).}

\item{d}{Numeric vector. Tree diameters (in centimeters).}

\item{ps}{Numeric variable. Plot area (in square meters).}
}
\value{
Numeric value representing the dominant tree height, which is
        the mean height of the 100 thickest trees.
}
\description{
This function calculates the dominant tree height, which is the mean height
of the 100 thickest trees in a specified plot area, based on their diameters.
}
\examples{
h <- c(10, 20, 30, 40, 50, 25, 30, 22, 11, 10)
d <- c(12, 25, 40, 40, 55, 35, 35, 25, 12, 15)
ps <- 400
h_dom(h, d, ps)

}
