% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivi_index_table.R
\name{ivi_index_table}
\alias{ivi_index_table}
\title{Calculate IVI Index Table}
\usage{
ivi_index_table(sp, xha, plot, data, plot_area)
}
\arguments{
\item{sp}{A column in `data` indicating species names.}

\item{xha}{A numeric column used for dominance calculation (e.g., basal area or crown area).}

\item{plot}{A factor or character column indicating plot identifiers.}

\item{data}{A data frame containing all the input columns.}

\item{plot_area}{Numeric value representing the area of the individual plots (e.g., 10000 m²).}
}
\value{
A data frame with species-wise IVI metrics: Abundance, Dominance, Frequency, and IVI.
}
\description{
Computes a table of Importance Value Index (IVI) for tree species based on
abundance (Nha), dominance (Xha), and frequency, using plot-level data. It will
compute the species frequency based on the number of plots. So it is recommended to compute the IVI
per area and not just for individual plots.
}
\examples{
# ivi_index_table(sp, xha, plot, data = mydata, plot_area = 400)
}
