% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margalef.R
\name{margalef}
\alias{margalef}
\title{Margalef's Species Richness Index}
\usage{
margalef(sp)
}
\arguments{
\item{sp}{A numeric or factor variable representing the species or individuals
in a sample.}
}
\value{
A numeric value representing Margalef's species richness index.
}
\description{
This function calculates Margalef's species richness index, which is an indicator
of the species richness in a given sample. The index is based on the number of
species and the total number of individuals. It is calculated using the following formula:
}
\details{
\deqn{D = \frac{S - 1}{\log(N)}}

where:
\itemize{
\item \eqn{S} is the number of unique species in the sample,
\item \eqn{N} is the total number of individuals in the sample.
}
}
\examples{
x <- c("P. patula", "A. religiosa", "A. religiosa", "Q. castanea", "P. patula")
margalef(x)

}
