% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/menhinick.R
\name{menhinick}
\alias{menhinick}
\title{Menhinick's Species Richness Index}
\usage{
menhinick(sp)
}
\arguments{
\item{sp}{A numeric or factor variable representing the species or individuals
in a sample.}
}
\value{
A numeric value representing Menhinick's species richness index.
}
\description{
This function calculates Menhinick's species richness index, which is another measure
of species diversity in a given sample. The index is calculated using the following formula:
}
\details{
\deqn{D = \frac{S}{\sqrt{N}}}

where:
\itemize{
\item \eqn{S} is the number of unique species in the sample,
\item \eqn{N} is the total number of individuals in the sample.
}
}
\examples{
x <- c("P. patula", "A. religiosa", "A. religiosa", "Q. castanea", "P. patula")
menhinick(x)

}
