% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nha.R
\name{Nha}
\alias{Nha}
\title{Number of Trees per Hectare}
\usage{
Nha(d, ps)
}
\arguments{
\item{d}{A numeric or character vector representing tree diameter or other tree attribute.}

\item{ps}{A numeric variable representing the plot area in square meters.}
}
\value{
A numeric value representing the number of trees per hectare.
}
\description{
This function calculates the number of trees per hectare based on the total
number of trees in a given plot area. The index is calculated using the following formula:
}
\details{
\deqn{N = \frac{N_{\text{trees}} \times 10000}{A}}

where:
\itemize{
\item \eqn{N_{\text{trees}}} is the total number of trees in the sample,
\item \eqn{A} is the plot area in square meters.
}
}
\examples{
x <- c(10, 20, 30, 40, 50, 25, 30, 22, 11, 10)
y <- 1000
Nha(x, y)

}
