% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnss_circle.R
\name{nnss_circle}
\alias{nnss_circle}
\title{Calculate Nearest Neighbor Indices for Circular Plots}
\usage{
nnss_circle(x, y, sp, d, h, r, data = NULL, center_point = NULL)
}
\arguments{
\item{x}{A numeric vector of x-coordinates for each tree.}

\item{y}{A numeric vector of y-coordinates for each tree.}

\item{sp}{A factor or character vector representing tree species.}

\item{d}{A numeric vector representing tree diameters.}

\item{h}{A numeric vector representing tree heights.}

\item{r}{A numeric value representing the radius of the circular plot (must be positive).}

\item{data}{A data frame containing all the required columns (\code{plot}, \code{x}, \code{y}, \code{sp}, \code{d}, \code{h}).}

\item{center_point}{Optional numeric vector including the center coordinates of the plots (cbind(17.84,17.84)).}
}
\value{
A data frame with calculated nearest neighbor indices for each tree.
Columns include:
\itemize{
\item \code{Ui}: Uniformity index
\item \code{Mi}: Mixture index
\item \code{dDomi}: Dominance based on diameter
\item \code{hDomi}: Dominance based on height
\item \code{dDif}: Diameter differentiation
\item \code{hDif}: Height differentiation
\item \code{NN1}: Nearest-neighbor edge-correction
}
}
\description{
This function calculates nearest neighbor indices for trees within a circular plot.
It requires coordinates, tree diameter, height, and species information, and returns
several indices related to tree competition and spacing.
}
\examples{

data("pipse_azimuth")

require(dplyr)

pipse_xy <-
pipse_azimuth \%>\%
mutate(coord_xy(azi = azimuth,dis = distance, r = 17.84*2))

df_nnss <- nnss_circle(x = x, y = y, sp = sp, d = d, h = h, r = 17.84, data = pipse_xy)
head(df_nnss)

}
