% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnss_square.R
\name{nnss_square}
\alias{nnss_square}
\title{Calculate Nearest Neighbor Indices for Square Plots}
\usage{
nnss_square(x, y, sp, d, h, xr, yr, data = NULL)
}
\arguments{
\item{x}{A numeric vector of x-coordinates for each tree.}

\item{y}{A numeric vector of y-coordinates for each tree.}

\item{sp}{A factor or character vector representing tree species.}

\item{d}{A numeric vector representing tree diameters.}

\item{h}{A numeric vector representing tree heights.}

\item{xr}{A numeric vector including the min and the max values of the x-coordinates for the plot.}

\item{yr}{A numeric vector including the min and the max values of the x-coordinates for the plot.}

\item{data}{A data frame containing all the required columns (\code{x}, \code{y}, \code{sp}, \code{d}, \code{h}).}
}
\value{
A data frame with calculated nearest neighbor indices for each tree.
Columns include:
\itemize{
\item \code{Ui}: Uniformity index
\item \code{Mi}: Mixture index
\item \code{dDomi}: Dominance based on diameter
\item \code{hDomi}: Dominance based on height
\item \code{dDif}: Diameter differentiation
\item \code{hDif}: Height differentiation
\item \code{NN1}: Nearest-neighbor edge-correction
}
}
\description{
This function calculates nearest neighbor indices for trees within a square plot.
It requires coordinates, tree diameter, height, and species information, along with the
plot's maximum x and y dimensions. The function returns several indices related to tree competition and spacing.
}
\examples{

data("pipse_one")

nnss_test <- nnss_square(x = x, y = y, sp = sp, d = d,
h = h, xr = cbind(0,50), yr = cbind(0,50), data = pipse_one)

head(nnss_test)

}
