% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pielou.R
\name{pielou}
\alias{pielou}
\title{Calculate Pielou's Equitativity Index}
\usage{
pielou(x)
}
\arguments{
\item{x}{Numeric or factor variable. A vector representing species or categories.}
}
\value{
Numeric. The Pielou equitativity index.
}
\description{
This function calculates Pielou's equitativity index, which measures the evenness
of species distribution in a community. The formula is:
\deqn{J' = \frac{H'}{\log(S)}}
where:
\itemize{
\item \eqn{H'} is the Shannon diversity index.
\item \eqn{S} is the number of unique categories or species.
}
}
\examples{
x <- c("n", "m", "n", "n", "m")
pielou(x)

}
