% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipse_azimuth.R
\docType{data}
\name{pipse_azimuth}
\alias{pipse_azimuth}
\title{Azimuth and distance data for Pinus pseudostrobus}
\format{
A data frame with 100 rows and 6 variables:
\describe{
  \item{plot}{Plot identifier (character)}
  \item{sp}{Species code, here `PIPSE` for *Pinus pseudostrobus* (character)}
  \item{d}{Diameter at breast height in cm (numeric)}
  \item{h}{Tree height in m (numeric)}
  \item{azimuth}{Azimuth from plot center to tree in degrees (numeric, 0–360)}
  \item{distance}{Distance from plot center to tree in meters (numeric)}
}
}
\source{
Field data collected by Ernesto Rubio.
}
\usage{
pipse_azimuth
}
\description{
This dataset includes tree measurements and spatial orientation information
for *Pinus pseudostrobus* individuals sampled in multiple plots.
}
\details{
The data were collected to analyze spatial distribution patterns and size variation of *Pinus pseudostrobus* in sampled forest plots. This dataset can be used for modeling spatial structure or basal area estimations.
}
\keyword{datasets}
