% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipse_one.R
\docType{data}
\name{pipse_one}
\alias{pipse_one}
\title{Single square plot for Pinus pseudostrobus}
\format{
A data frame with 127 rows and 5 variables:
\describe{
  \item{x}{Unique identifier for the sample plot (character)}
  \item{y}{Species code (character)}
  \item{sp}{Species code (character)}
  \item{d}{Diameter at breast height in cm (numeric)}
  \item{h}{Tree total height in m (numeric)}
}
}
\source{
Field data collected by Ernesto Rubio.
}
\usage{
pipse_one
}
\description{
This dataset contains measurements collected in a square plots established
for the study of *Pinus pseudostrobus* stands.
}
\details{
The data were collected in square plots following standard forest
inventory methodology. They can be used for spatial structure analysis,
stand density estimation, or modeling of tree growth.
}
\keyword{datasets}
