% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipse_square.R
\docType{data}
\name{pipse_square}
\alias{pipse_square}
\title{Square plot data for Pinus pseudostrobus}
\format{
A data frame with 4260 rows and 9 variables:
\describe{
  \item{x}{x coordinate for the tree}
  \item{y}{y coordinate for the tree}
  \item{sp}{Species code (character)}
  \item{d}{Diameter at breast height in cm (numeric)}
  \item{h}{Tree total height in m (numeric)}
  \item{ca}{Crown area}
  \item{x_c}{x coordinate for the plot center}
  \item{y_c}{y coordinate for the plot center}
  \item{plot}{Unique identifier for the sample plot}
}
}
\source{
Field data collected by Ernesto Rubio.
}
\usage{
pipse_square
}
\description{
This dataset contains measurements collected in square plots (50x50 m) established
for the study of *Pinus pseudostrobus* stands.
}
\details{
The data were resampled from the original data in square plots (50x50 meters)
following standard forest inventory methodology. They can be used for spatial structure analysis,
stand diversity, and structure estimation, or modeling.
}
\keyword{datasets}
