% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q1.R
\name{q1}
\alias{q1}
\title{Hill's number for q1}
\usage{
q1(sp)
}
\arguments{
\item{sp}{A numeric or factor vector representing categories (e.g., species or groups).}
}
\value{
A numeric value representing Hill's number for q1 index.
}
\description{
This function calculates exp value for the Shannon's entropy index (\eqn{H'}) for a given dataset.
}
\details{
The formula is:
\deqn{q1 = exp(- \sum_{i=1}^S p_i \log(p_i))}

Where:
\itemize{
\item \eqn{p_i} is the proportion of individuals belonging to the \eqn{i}-th category.
\item \eqn{S} is the total number of unique categories in the dataset.
}
}
\examples{
x <- c("n", "m", "n", "n", "m")
q1(x)

}
