% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2.R
\name{q2}
\alias{q2}
\title{Hill's number for q2 based on the Simpson's index.}
\usage{
q2(sp)
}
\arguments{
\item{sp}{Numeric or factor variable}
}
\value{
Numeric variable.
}
\description{
Hill's number for q2 based on the Simpson's index.
}
\examples{

x <- c("n", "m", "n", "n", "m")

q2(x)

}
