% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.R
\name{rich}
\alias{rich}
\title{Calculate Species Richness (S)}
\usage{
rich(x)
}
\arguments{
\item{x}{Numeric or factor variable. A vector representing species or categories.}
}
\value{
Numeric. The species richness (\eqn{S}).
}
\description{
This function calculates species richness (\eqn{S}), which represents the
number of unique species or categories in the given data. The formula is:
\deqn{S = \sum I(counts > 0)}
where:
\itemize{
\item \eqn{counts} represents the frequency of each unique species or category.
}
}
\examples{
x <- c("n", "m", "n", "n", "m")
rich(x)

}
