% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdi.R
\name{sdi}
\alias{sdi}
\title{Calculate Stand Density Index (SDI)}
\usage{
sdi(x, y)
}
\arguments{
\item{x}{Numeric vector. A vector of tree diameters (e.g., in cm).}

\item{y}{Numeric. The plot area (e.g., in square meters).}
}
\value{
Numeric. The stand density index.
}
\description{
This function calculates the Stand Density Index (SDI) using the formula:
\deqn{SDI = Nha \times \left(\frac{dg}{25.4}\right)^{1.605}}
where:
\itemize{
\item \eqn{Nha} is the number of trees per hectare.
\item \eqn{dg} is the quadratic mean diameter (in cm).
}
}
\examples{
x <- c(10, 20, 30, 40, 50, 25, 30, 22, 11, 10)
y <- 1000
sdi(x, y)

}
