% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{Shannon's Entropy Index}
\usage{
shannon(x)
}
\arguments{
\item{x}{A numeric or factor vector representing categories (e.g., species or groups).}
}
\value{
A numeric value representing Shannon's entropy index (\eqn{H'}).
}
\description{
This function calculates Shannon's entropy index (\eqn{H'}) for a given dataset.
Shannon's entropy index measures the diversity in the dataset, considering both
the richness (number of unique categories) and evenness (distribution of individuals).
}
\details{
The formula is:
\deqn{H' = - \sum_{i=1}^S p_i \log(p_i)}

Where:
\itemize{
\item \eqn{p_i} is the proportion of individuals belonging to the \eqn{i}-th category.
\item \eqn{S} is the total number of unique categories in the dataset.
}
}
\examples{
x <- c("n", "m", "n", "n", "m")
shannon(x)

}
