% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon_evenness.R
\name{shannon_evenness}
\alias{shannon_evenness}
\title{Shannon's Evenness Index}
\usage{
shannon_evenness(x)
}
\arguments{
\item{x}{A numeric or factor vector representing species or groups in the community.}
}
\value{
A numeric value representing Shannon's evenness index, which ranges from 0 (uneven) to 1 (perfectly even).
}
\description{
This function calculates Shannon's evenness index, which measures the
evenness of the distribution of species (or any type of data) in a community.
The formula used is:
}
\details{
\deqn{E_H = \frac{H'}{H_{\text{max}}}}

where:
\itemize{
\item \eqn{H'} is the Shannon diversity index:
\deqn{H' = - \sum p_i \log(p_i)}
with \eqn{p_i} being the proportion of individuals of species \eqn{i}.
\item \eqn{H_{\text{max}}} is the maximum Shannon index for a given number of unique species \eqn{S},
and is given by \eqn{H_{\text{max}} = \log(S)}.
}
}
\examples{
x <- c("n", "m", "n", "n", "m")
shannon_evenness(x)

}
