% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson.R
\name{simpson}
\alias{simpson}
\title{Simpson's index.}
\usage{
simpson(sp)
}
\arguments{
\item{sp}{Numeric or factor variable}
}
\value{
Numeric variable.
}
\description{
Simpson's index.
}
\examples{

x <- c("n", "m", "n", "n", "m")

simpson(x)
}
