% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpson_complement.R
\name{simpson_complement}
\alias{simpson_complement}
\title{Simpson's Index Complement}
\usage{
simpson_complement(sp)
}
\arguments{
\item{sp}{A numeric or factor vector representing categories (e.g., species or groups).}
}
\value{
A numeric value representing the complement of Simpson's index (\eqn{D'}).
}
\description{
This function calculates the complement of Simpson's index (\eqn{D'}), a diversity measure
that represents the probability of two randomly selected individuals belonging to
different categories.
}
\details{
The formula is:
\deqn{D' = 1 - \sum_{i=1}^S p_i^2}

Where:
\itemize{
\item \eqn{p_i} is the proportion of individuals in the \eqn{i}-th category,
\item \eqn{S} is the total number of categories.
}
}
\examples{
x <- c("n", "m", "n", "n", "m")
simpson_complement(x)

}
