% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A_rel.R
\name{A_rel}
\alias{A_rel}
\title{Relative Value for the Species Profile Index A}
\usage{
A_rel(x, y)
}
\arguments{
\item{x}{A numeric vector representing tree heights.}

\item{y}{A character or factor vector representing species.}
}
\value{
A numeric value representing the relative Species Profile Index A (as a percentage).
}
\description{
This function calculates the relative value of the Species Profile Index A,
which represents the observed species diversity in relation to the maximum
possible diversity for a given set of species. The relative value is expressed
as a percentage.
}
\examples{
x <- c(10, 20, 30, 40, 50, 25, 30, 22, 11, 10)
y <- c(
  "P. patula", "P. patula", "P. patula", "A. religiosa",
  "Q. castanea", "P. patula", "P. patula", "P. patula",
  "P. patula", "P. patula"
)
A_rel(x, y)

}
