% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnss_circle.R
\name{nnss_circle}
\alias{nnss_circle}
\title{Calculate Nearest Neighbor Indices for Circular Plots}
\usage{
nnss_circle(plot, x, y, sp, d, h, r, data = NULL)
}
\arguments{
\item{plot}{A numeric or character variable representing the plot identifier.}

\item{x}{A numeric vector of x-coordinates for each tree.}

\item{y}{A numeric vector of y-coordinates for each tree.}

\item{sp}{A factor or character vector representing tree species.}

\item{d}{A numeric vector representing tree diameters.}

\item{h}{A numeric vector representing tree heights.}

\item{r}{A numeric value representing the radius of the circular plot (must be positive).}

\item{data}{A data frame containing all the required columns (\code{plot}, \code{x}, \code{y}, \code{sp}, \code{d}, \code{h}).}
}
\value{
A data frame with calculated nearest neighbor indices for each tree.
Columns include:
\itemize{
\item \code{Ui}: Uniformity index
\item \code{Mi}: Mixture index
\item \code{dDomi}: Dominance based on diameter
\item \code{hDomi}: Dominance based on height
\item \code{dDif}: Diameter differentiation
\item \code{hDif}: Height differentiation
\item \code{NN1}: Nearest-neighbor edge-correction
}
}
\description{
This function calculates nearest neighbor indices for trees within a circular plot.
It requires coordinates, tree diameter, height, and species information, and returns
several indices related to tree competition and spacing.
}
\examples{
plot_radius <- 17.84124 # for a 1000 m2 plot
azi <- runif(100, min = 0, max = 360)
dis <- runif(100, min = 1, max = plot_radius)
df_xy <- coordxy(azi, dis, plot_radius*2)
sp <- factor(sample(c("Pinus", "Quercus"), 100, replace = TRUE))
d <- runif(100, min = 7.5, max = 60)
h <- runif(100, min = 2, max = 40)
plot <- rep("P01", 100)
dataP02 <- data.frame(plot, df_xy, sp, d, h)
nnss_circle(plot = plot, x = x, y = y, sp = sp, d = d, h = h, r = plot_radius, data = dataP02)

}
