% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnss_square.R
\name{nnss_square}
\alias{nnss_square}
\title{Calculate Nearest Neighbor Indices for Square Plots}
\usage{
nnss_square(plot, x, y, sp, d, h, xmax, ymax, data = NULL)
}
\arguments{
\item{plot}{A numeric or character variable representing the plot identifier.}

\item{x}{A numeric vector of x-coordinates for each tree.}

\item{y}{A numeric vector of y-coordinates for each tree.}

\item{sp}{A factor or character vector representing tree species.}

\item{d}{A numeric vector representing tree diameters.}

\item{h}{A numeric vector representing tree heights.}

\item{xmax}{A numeric value representing the maximum x-coordinate value for the plot.}

\item{ymax}{A numeric value representing the maximum y-coordinate value for the plot.}

\item{data}{A data frame containing all the required columns (\code{plot}, \code{x}, \code{y}, \code{sp}, \code{d}, \code{h}).}
}
\value{
A data frame with calculated nearest neighbor indices for each tree.
Columns include:
\itemize{
\item \code{Ui}: Uniformity index
\item \code{Mi}: Mixture index
\item \code{dDomi}: Dominance based on diameter
\item \code{hDomi}: Dominance based on height
\item \code{dDif}: Diameter difference
\item \code{hDif}: Height difference
\item \code{NN1}: Distance to the nearest neighbor
}
}
\description{
This function calculates nearest neighbor indices for trees within a square plot.
It requires coordinates, tree diameter, height, and species information, along with the
plot's maximum x and y dimensions. The function returns several indices related to tree competition and spacing.
}
\examples{
set.seed(42)
x <- runif(100, min = 1, max = 50)
y <- runif(100, min = 1, max = 50)
sp <- factor(sample(c("Pinus", "Quercus"), 100, replace = TRUE))
d <- runif(100, min = 7.5, max = 60)
h <- runif(100, min = 2, max = 40)
plot <- rep("P01", 100)
dataP01 <- data.frame(plot, x, y, sp, d, h)
nnss_square(plot = plot, x = x, y = y, sp = sp, d = d, h = h, xmax = 50, ymax = 50, data = dataP01)

}
